Move more logic
