+++ ert3/engine/__init__.py
@@ -1,3 +1,2 @@
 from ert3.engine._run import run
 from ert3.engine._export import export
-from ert3.engine._init import init_workspace
