+++ ert3/engine/_export.py
@@ -6,9 +6,9 @@ from pathlib import Path
 
 def export(workspace_root, experiment_name):
     experiment_root = Path(workspace_root) / experiment_name
-    ert3._assert_experiment(workspace_root, experiment_name)
+    ert3.workspace.experiment_exists(workspace_root, experiment_name)
 
-    if not ert3._experiment_have_run(workspace_root, experiment_name):
+    if not ert3.workspace.experiment_have_run(workspace_root, experiment_name):
         raise ValueError("Cannot export experiment that has not been carried out")
 
     input_data = ert3.storage.get_input_data(workspace_root, experiment_name)
