+++ ert3/workspace/__init__.py
@@ -1,0 +1,4 @@
+from ert3.workspace._workspace import locate_root
+from ert3.workspace._workspace import experiment_exists
+from ert3.workspace._workspace import experiment_have_run
+from ert3.workspace._workspace import init
